# Optimize decoding of keys and packed fields